requirements.txt
asttokens==2.2.1
backcall==0.2.0
black==22.12.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
flake8==6.0.0
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-translate==3.9.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
ipykernel==6.19.2
ipython==8.7.0
jedi==0.18.2
jupyter-client==7.4.8
jupyter-core==5.1.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==0.4.3
nest-asyncio==1.5.6
packaging==22.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.0
prompt-toolkit==3.0.36
proto-plus==1.22.1
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.13.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pyzmq==24.0.1
requests==2.28.1
rsa==4.9
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.7.1
typing-extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
